requirements.txt
daft[deltalake]
python-dotenv